refactor: separate out fluid utils in files
