+++ src/libecalc/domain/process/entities/fluid_stream/__init__.py
@@ -1,8 +1,9 @@
 from libecalc.domain.process.entities.fluid_stream.conditions import ProcessConditions
+from libecalc.domain.process.entities.fluid_stream.eos_model import EoSModel
+from libecalc.domain.process.entities.fluid_stream.fluid_composition import FluidComposition
 from libecalc.domain.process.entities.fluid_stream.fluid_stream import FluidStream
 from libecalc.domain.process.entities.fluid_stream.mixing import SimplifiedStreamMixing
 from libecalc.domain.process.entities.fluid_stream.thermo_system import ThermoSystemInterface
-from libecalc.domain.process.entities.fluid_stream.utils import EoSModel, FluidComposition
 
 __all__ = [
     "FluidStream",
