+++ src/libecalc/domain/process/entities/fluid_stream/eos_model.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class EoSModel(str, Enum):
+    SRK = "SRK"
+    PR = "PR"
+    GERG_SRK = "GERG_SRK"
+    GERG_PR = "GERG_PR"
