+++ src/libecalc/presentation/yaml/mappers/fluid_mapper.py
@@ -1,7 +1,8 @@
 from typing import Any
 
 from libecalc.common.fluid import FluidModel
-from libecalc.domain.process.entities.fluid_stream.utils import EoSModel, FluidComposition
+from libecalc.domain.process.entities.fluid_stream.eos_model import EoSModel
+from libecalc.domain.process.entities.fluid_stream.fluid_composition import FluidComposition
 from libecalc.domain.resource import Resources
 from libecalc.presentation.yaml.yaml_keywords import EcalcYamlKeywords
 from libecalc.presentation.yaml.yaml_types.models import YamlFluidModel
