+++ tests/ecalc_neqsim_wrapper/conftest.py
@@ -2,7 +2,8 @@ import pytest
 
 from ecalc_neqsim_wrapper import NeqsimService
 from ecalc_neqsim_wrapper.thermo import NeqsimFluid
-from libecalc.domain.process.entities.fluid_stream.utils import EoSModel, FluidComposition
+from libecalc.domain.process.entities.fluid_stream.eos_model import EoSModel
+from libecalc.domain.process.entities.fluid_stream.fluid_composition import FluidComposition
 
 HEAVY_FLUID_COMPOSITION = FluidComposition(
     nitrogen=0.682785869,
