+++ tests/ecalc_neqsim_wrapper/unit_tests/test_neqsim_fluid.py
@@ -5,7 +5,8 @@ import pytest
 
 from ecalc_neqsim_wrapper.exceptions import NeqsimComponentError
 from ecalc_neqsim_wrapper.thermo import NeqsimFluid, mix_neqsim_streams
-from libecalc.domain.process.entities.fluid_stream.utils import EoSModel, FluidComposition
+from libecalc.domain.process.entities.fluid_stream.eos_model import EoSModel
+from libecalc.domain.process.entities.fluid_stream.fluid_composition import FluidComposition
 from tests.ecalc_neqsim_wrapper.conftest import MEDIUM_MW_19P4_COMPOSITION
