+++ tests/libecalc/domain/core/stream/test_mixing.py
@@ -1,13 +1,14 @@
 import pytest
 
 from libecalc.domain.process.entities.fluid_stream.conditions import ProcessConditions
+from libecalc.domain.process.entities.fluid_stream.eos_model import EoSModel
 from libecalc.domain.process.entities.fluid_stream.exceptions import (
     IncompatibleEoSModelsException,
     IncompatibleThermoSystemProvidersException,
 )
+from libecalc.domain.process.entities.fluid_stream.fluid_composition import FluidComposition
 from libecalc.domain.process.entities.fluid_stream.fluid_stream import FluidStream
 from libecalc.domain.process.entities.fluid_stream.mixing import SimplifiedStreamMixing
-from libecalc.domain.process.entities.fluid_stream.utils import EoSModel, FluidComposition
 from libecalc.infrastructure.thermo_system_providers.neqsim_thermo_system import NeqSimThermoSystem
