+++ tests/libecalc/domain/core/stream/test_thermo_system.py
@@ -4,7 +4,7 @@ import pytest
 
 from ecalc_neqsim_wrapper.thermo import NeqsimFluid
 from libecalc.domain.process.entities.fluid_stream.conditions import ProcessConditions
-from libecalc.domain.process.entities.fluid_stream.utils import EoSModel
+from libecalc.domain.process.entities.fluid_stream.eos_model import EoSModel
 from libecalc.infrastructure.thermo_system_providers.neqsim_thermo_system import NeqSimThermoSystem
