+++ tests/libecalc/dto/test_composition.py
@@ -1,7 +1,7 @@
 import pytest
 from pydantic import ValidationError
 
-from libecalc.domain.process.entities.fluid_stream.utils import FluidComposition
+from libecalc.domain.process.entities.fluid_stream.fluid_composition import FluidComposition
 
 
 def test_composition(caplog):
